Move MW work to subdirectory
