+++ components/MW/metaboliteInfo/package.json
